requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=14.0
datasets>=2.16
huggingface-hub>=0.20
requests>=2.31
beautifulsoup4>=4.12
feedparser>=6.0
matplotlib>=3.8
pyyaml>=6.0
python-dotenv>=1.0
kaggle>=1.6
nbformat>=5.9
nbconvert>=7.0
ipykernel>=6.0
anthropic>=0.40
scikit-learn>=1.3
joblib>=1.3
label-studio>=1.12
label-studio-sdk>=1.0
